+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+2.2.0
+-----
+
+- Add option `config.length` for support change number chars.
+- Add option `config.strikethrough` for enable or disable strikethrough.
+
 2.1.3
 -----